requirements.txt
c2pa-python
cryptography
numpy
opencv-python
pillow
protobuf
pyzbar
qrcode
streamlit
